[print] Move print function to bndlib
Printing of a JAR is done in several places. This moves the printing to bndlib
so it can be used anywhere.
