+++ biz.aQute.bndlib/src/aQute/bnd/print/package-info.java
@@ -1,0 +1,6 @@
+@Export
+@Version("1.0.0")
+package aQute.bnd.print;
+
+import org.osgi.annotation.bundle.Export;
+import org.osgi.annotation.versioning.Version;
